requirements.txt
telebot
pygments
pillow
